requirements.txt
sentence-transformers
faiss-cpu
networkx
pandas
numpy
python-dotenv
streamlit
openai
tqdm
